Updated resolve condition only if backend connection fails
The current resolve condition is based on return anything other than 200 (OK). This is sub-optimal as it will keep trying even if it reached the desired application. Possible return codes:
- 201: if a resource was created
- 404: if a resource was not found.

As commented in this issue (https://github.com/Azure/service-fabric-issues/issues/346), a better starting point would be resolving only if last error was backend connection failure.